Perubahan Calculator pada tipe data

+++ Calculator.cs
@@ -53,17 +53,17 @@
             return a + b;
         }
         
-        static int Pengurangan(int a, int b)
+        static double Pengurangan(int a, int b)
         {
             return a - b;
         }
         
-        static int Perkalian(int a, int b)
+        static double Perkalian(int a, int b)
         {
             return a * b;
         }
 
-        static int Pembagian(int a, int b)
+        static double Pembagian(int a, int b)
         {
             return a / b;
         }
